refactor(baseline): more cmd restructuring
